Bring in jQuery via CDN
Minifying jQuery with the other dependencies is known to
cause odd issues in earlier versions of IE. Furthermore,
with so many dependencies relying on jQuery, we need to
ensure it is loaded by the time the scripts in the footer
get parsed. Although not ideal, this seemed liked the least hassle.