Ensure DateTimeField carries over DB value
Not entirely sure what was causing this behaviour but I
believe it was caused by the test data creating DateTime
instances containing seconds greater than zero, which
could result in slightly incorrect dates being displayed
on the FE. Perhaps to do with the library being configured
not to deal with seconds?

Not really happy with the solution but something better
doesn't spring to mind.