Remove "stepMinutes" usage from DateTimeField
Ensuring the selection always starts on an increment
of five is proving rather tricky. It gets messy with dates
such as 31/12/2014 23:59:00 as adding time results in a lot
of rolling over which javascript's Date object doesn't handle.
Could be achieved easily with a third-party library but seeing
as no client has explicitly requested this behaviour so far
I shall undo.